[FUA-98] Turn <PageFooter> into a wrapper
