+++ src/renderer/components/PageFooter/styles.pcss
@@ -10,8 +10,3 @@
     position: fixed;
     width: 100vw;
 }
-
-.footer-button {
-    margin-right: 15px;
-    width: 200px;
-}
