+++ src/renderer/containers/UploadSelectionPage/styles.pcss
@@ -32,6 +32,11 @@
   flex-direction: column;
 }
 
+.footer-button {
+    margin-right: 15px;
+    width: 200px;
+}
+
 .spin-container {
   height: 100%;
   width: 100%;
